Cleanups to the openai completion client
