+++ internal/completions/client/openai/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     srcs = [
         "decoder.go",
         "openai.go",
+        "types.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/completions/client/openai",
     tags = [TAG_CODY_CORE],
